feat: intermediate update porting

+++ src/osu/difficulty/scaling_factor.rs
@@ -23,7 +23,7 @@ impl ScalingFactor {
             * BROKEN_GAMEFIELD_ROUNDING_ALLOWANCE;
 
         let radius = f64::from(OsuObject::OBJECT_RADIUS * scale);
-        let factor = OsuDifficultyObject::NORMALIZED_RADIUS / radius as f32;
+        let factor = OsuDifficultyObject::NORMALIZED_RADIUS as f32 / radius as f32;
 
         let factor_with_small_circle_bonus = if radius < 30.0 {
             factor * (1.0 + (30.0 - radius as f32).min(5.0) / 50.0)
